Use sampling frequency reported by underlying amp
instead of calculating it on every round. We also check if markers are
younger or older than the current block.

If the marker is younger, it is queued for the next block by putting it
back into the marker_queue without losing precision. If it is older we
have to count it to the 0th sample of the current block. We mitigate the
second case which leads to loss of information by waiting .2ms after
getting the data from the amp for incoming markers.